pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.0"
python = "^3.8"
drf-yasg = "1.21.5"
django-rest-framework = "0.1.0"
django-user = "^0.2.1"
[tool.poetry.dev-dependencies]
